requirement.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.7.1
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
phonenumbers==8.13.33
pycodestyle==2.11.1
pydantic==1.10.14
pyflakes==3.2.0
PyJWT==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==1.4.36
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==12.0
